--- src/algorithms/index.ts
@@ -1,21 +1,0 @@
-import { __BubbleSort as BubbleSort } from './bubble_sort';
-import { __CocktailShakerSort as CocktailShakerSort } from './cocktail_shaker';
-import { __InsertionSort as InsertionSort } from './insertion_sort';
-import { __SelectionSort as SelectionSort } from './selection_sort';
-import { __QuickSort as QuickSort } from './quick_sort';
-
-export const sorts = [
-  BubbleSort,
-  CocktailShakerSort,
-  SelectionSort,
-  InsertionSort,
-  QuickSort,
-] as const;
-
-export {
-  BubbleSort,
-  CocktailShakerSort,
-  SelectionSort,
-  InsertionSort,
-  QuickSort,
-};
