+++ src/entities/algorithms/index.ts
@@ -1,0 +1,3 @@
+import { SortHistory } from 'entities/sort_history';
+
+type Algorithm = (number: []) => SortHistory[];
