+++ src/features/sort_array/model/cocktail_shaker/index.ts
@@ -5,9 +5,9 @@ import {
   addToSwapping,
   cleanStatuses,
   initializeSteps,
-} from './helpers';
+} from '../../lib';
 
-export const __CocktailShakerSort = (array: number[]) => {
+export const CocktailShakerSort = (array: number[]) => {
   let nums = array.slice();
   let historySteps = initializeSteps(nums);
